Update patch set 2
Patch Set 2: Fails; No score

Unfortunately this change doesn't merge cleanly on the latest head (it looks like it was originally made in a cupcake timeframe and the underlying code has changed).

If it's still relevant, please resolve the conflicts and upload a replacement. Thanks.

Patch-set: 2
Label: Code-Review=0
Label: Verified=-1