Update patch set 2
Patch Set 2: No score

Hi, David!

Thanks for reviewing, I'll try explaining below.

It makes the code a little "pessimistic", by testing for equality only if it doesn't branch to the "right". 

It doesn't makes the "left" case worse, as 2 tests will be still necessary (not equals and also not branching "right").

For the trivial case of 1 element, it makes it worse because two tests will be necessary instead of just one.

For the case of searching only on the "left" branch, it adds just one extra comparison (the last one, before testing for element equality).

If we take just one branch to the "right", it avoids a comparison every time we go "right" and introduces a comparison (the last one, before testing for element equality).

But it is better when we navigate "right" (roughly ~50% of the searchs) more than once because it avoids testing for equality every time we are going to go to the "right".

For any reasonable big arrays, this means that:

a. when branching "left", two tests are still needed (no change from original code);

b. when branching "right", only one test will be needed (from the two which were originally necessary whenever navigating to the "right") on every iteration;

c. when branching only left or on the trivial case with one element, we introduce one (and only one) extra comparison;

Most of bsearch() implementations I've came across also optimizes for reasonably sized arrays, because it makes the slower case (big arrays) faster, without making the small arrays case necessarily slower.

In my opinion it's a good trade-off (glibc also does the same). In 2007, I've opened an improvement request to SUN on java/util/Arrays.java with this same suggestion and it was accepted and incorporated into its code at that time.

Patch-set: 2
Reviewer: André Goddard Rosa <1003237@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0