Update patch set 1
Patch Set 1:

what I mean is that the __POSIX_VISIBLE macros are defined by the C library headers themselves (currently in <sys/limits.h>, but this should probably be moved to <sys/cdefs.h>).

Note the double leading underscore here, unlike the feature test macros like _POSIX_VISIBLE.

An alternative is to define the macro as part of LOCAL_CFLAGS in Android.mk since we want to use everything when building the C library.

Patch-set: 1