Update patch set 1
Patch Set 1: Do not submit

It is not acceptable to submit code on behalf of another copyright holder, without following the process described in the CLA for submitting code on behalf of someone else. Beyond that, it is separately not acceptable to remove a copyright header from existing code.

This is a violation of the CLA you agreed to when you signed up at review.source.android.com. Please don't do this again.

Patch-set: 1
Reviewer: Dan Morrill <1000962@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Verified=0