Update patch set 1
Patch Set 1:

I meant to say:

In theory, an application wanting to modify the mapped region just has to call mprotect to change it from read-only to read-write.  As a result, the existing read-only setting doesn't give any security protections.

This change doesn't make the problem worse than it already is.

Patch-set: 1