Update patch set 1
Patch Set 1:

Regardless of whether apps *should* be able to modify the page, the new kernel futex implementation *requires* read/write access to futex cells.  It's an hack as far as I'm concerned (as the futex call doesn't modify the cell's value), but that's the way it is.

So probably to fix this "right" would require some type of client/server thing where apps request the property value via some IPC mechanism.

I think the existing code (prior to this change) allows a writable mmap() because (I assume) the fd that the /dev/ashmem section has write access.

Patch-set: 1