Update patch set 1
Patch Set 1: Fails

Unfortunately this conflicts with the latest master.

Taking a glance at the history, it appears that this change was built on top of a "naked" froyo. However, the master branch contains several months worth of additional contributions, which must be the reason for the conflict.

Patch-set: 1
Label: Code-Review=0
Label: Verified=-1