Update patch set 3
Patch Set 3:

Here's a new version which splits out the logic to its own file, adds proper locking, and uses the queue structures.  It does not yet include logic to remove the handlers during dlclose()--I'm unclear whether you were saying you'd like to see that implemented before pushing this, or whether it's ok to leave it out for now since atexit() also does not implement this behavior.

I'm happy to implement this, but would like some suggestions on how to proceed, as it does indeed seem as if it will be a bit hairy.  Glibc appears to implement it by performing the unregister as part of __cxa_finalize(), which would work here too.  The only difficult part is registering the handler with a reference to the proper dso, so that we know which ones to clear out.  Glibc appears to do this with some really crazy stuff involving symbol hiding, to ensure that a different version of the function gets invoked for each library, with a reference to the proper object.

I'm not quite sure how that would be implemented here, but another thought I had was to use the same technique that I used to implement RTLD_NEXT in dlsym(), which was to use __builtin_return_address() to get the caller's address. Then that could be fed to dladdr() to get the dso reference, which could then be stored along with the handlers for use during unregistering.  That's somewhat cumbersome, but I think it would work--does anybody have any suggestions on whether there would be a simpler way to do this?

Patch-set: 3
Label: Code-Review=0