Update patch set 3
Patch Set 3:

On second thought, dladdr() won't return the correct dso reference, it just returns the library name.  But could a similar approach be taken, with some kind of secret linker function that the C library could call which would return the proper reference?  That would allow both the atfork stuff and the atexit stuff to register themselves with the proper dso pointers so that they can be unregistered during dlclose().

Patch-set: 3