Update patch set 1
Patch Set 1:

nice! thanks. i'm curious about the specific list of ARM variants, though; is __ARM_ARCH_7M__ missed out on purpose, for example?

i really think we should use cpu-features.h and __ARM_ARCH__ >=6, even if we have to hack around PLD, because otherwise we have to keep two lists up to date. worse, the default behavior will be to turn _off_ these optimizations.

if we're going to duplicate the list, how about reversing the sense, and listing the architectures we _don't_ have REV/REV16 on? that list should be smaller and unchanging...

Patch-set: 1