Update patch set 1
Patch Set 1:

Just as a note, we've had to revert this change in our internal master tree, as it somehow broke our automated testing. We're assuming that some test tool relies on signals not getting masked as they should and ends up getting a SIGHUP that it doesn't know how to deal with when we disconnect USB during testing.

Patch-set: 1