Update patch set 2
Patch Set 2:

Bruce, I understand this, but we also need to export an atexit() function from the C library to avoid breaking the loading of buggy application shared libraries.

I understand that the exit handler of these libraries was never called, but we shouldn't prevent them from loading due to this.

The function is part of the NDK ABI, and the code that uses it has already been generated out of our control. We need a way to cope with this case.

NOTE: This is only a problem for the ARM ABI, the x86 ABI is not public yet, so we should "just do the right thing" now for it.

Patch-set: 2