Update patch set 2
Patch Set 2:

I see your point. It's this problem that caused me to upload the change for external/stlport. It was using an old CRT* suite... and the linker gave a warning about dso_handle.

SO... if we use this implementation but didn't mark the libc version as hidden... what's the consequence for an existing .so file (and of course... all NDK apps have them... so there are quite a few)...

  o  Since foojni.so doesn't have it's own atexit routine, it will call the (not hidden) libc routine.
  o the libc routine will pass IT'S dsohandle to the cxa routine
  o Since libc is never unloaded, the foojni.so destructors will never be called.

If foojni.so is loaded/unloaded many times... we have a memory leak (ARM has this link now... if I understand correctly).

Do I have the above correct? We willing to live with it for ARM?

Patch-set: 2