Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

memcpy() of overlapping is undefined behaviour according to the standard, but not according to the current implementation (for ARM and x86), and I think it's a valuable optimization to keep for these.

If this code is not correct for you, I'd suggest instead to provide a version that suits your needs under arch-mips/bionic/ and ensure that Android.mk will pick it for your architecture, instead of this one.

Patch-set: 1
Label: Code-Review=-1