Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Ok, this is a non-trivial patch that modifies some public C library definitions with a bunch of #ifdef __mips__ which I feel are pretty awkward, but there are probably ways to work around that.

Before going any further, we would like to avoid repeating the mistakes we made when accepting the SuperH support (which was never maintained and bit-rotted heavily), which we are currently trying to remove.

We really need the following to support MIPS in any official capacity, and that's only at the technical level:

1/ A kernel tree that is capable of generating MIPS-specific kernel image, and that we can host on android.googlesource.com. Do the sources under android.googlesource.com/kernel/common.git would be enough? Maybe you already have an open-source kernel tree we could pull from?

It doesn't need to be a functioning hardware-specific kernel image, but at a minimum we need all the headers to be correct, so we can copy them to $ANDROID/external/kernel-headers/original and generate cleaned-up versions from them.

(For SuperH, we accepted stuff into external/kernel-headers without looking at it in more details. It turned out that these came from a never-released kernel tree and we have no way to know if they're still valid).

Ideally, this kernel should be as close as possible from he kernel/common one used by AOSP, but ymmv.

2/ A valid GCC cross-toolchain that works on Linux and  OS X, and that we can rebuild from the sources coming from a public repository. We highly favor using android.googlesource.com/toolchain/ to host the platform's toolchain sources, and we would probably want to put the MIPS sources there too. Ideally, using the exact same sources for all target architectures (we currently do it for ARM and x86).

I know there is a chicken-and-egg problem, because you'll need crtbegin/crtend object files to build the cross toolchain itself, and you won't have gdbserver until the C library works, but this one can be provided later.

3/ Scripts to rebuild the toolchain automatically, so we can put binaries under prebuilt/. JBQ can provide more information here. Any binaries uploaded here should have complete rebuild instructions in the commit message.

4/ Build system changes (mainly build/core/combo/ stuff) to support MIPS, including the ARCH_MIPS_HAVE_XXXX macro definitions used for device-specific features that can be tested at build time, to optimize the final binary for specific device build. Note the convention if ARCH_$ARCH_HAVE_XXX, not ARCH_$ARCH_HAS_XXX, as I've seen in a few patches.

5/ A build product that can be used to test MIPS builds automatically. E.g. we use full-eng and full_x86-eng already (they build emulator-specific images). You could consider a full_mips-eng product, even if the emulator doesn't support MIPS emulation, at least to check that nothing breaks.

It may be "empty" at first, but should grow as platform contributions are made progressively.

This will ensure that platform modifications that break the MIPS-dependent build could be caught reasonably soon. I can't guarantee that we're going to run the full_mips eng on our build servers for now though.

Once all these are setup, we'll be ready to accept Bionic and other contributions to support MIPS in a sane way.

I'll let others make additional remarks.

Patch-set: 1
Label: Code-Review=-1