Update patch set 1
Patch Set 1:

Kernel:

There are some MIPS kernel changes on top of android.googlesource.com/kernel/common.git

You can download the MIPS kernel sources with the below command:

$ git clone git://git.linux-mips.org/pub/scm/linux-mti.git -b dev-android-linux-mti-3.0.8

I also have patches for the goldfish kernel which I will submit to the below repository for the branch android-goldfish-2.6.29 (?).

https://android.googlesource.com/kernel/goldfish.git

Toolchain:

We currently use the compiler from CodeSourcery which is based on GCC 4.5.2 and is tested on linux only. This compiler has patches to support newer MIPS targets. We are currently hosting the compiler sources on Github at https://github.com/mipsgcc

The prebuilt compiler can be got from git://github.com/mipsgcc/mips-linux-android-4.5.2.git.  

The scripts to build the compiler are also ready and will be uploaded. Is this an acceptable form of submitting the compiler if we host our own compiler sources?

In the past, we had built compilers for OS X, so it should not be a problem to build one from the sources, although we are not using one currently.

Build system Changes:

We are using ARCH_$ARCH_HAS_XXX. I will modify them to use ARCH_$ARCH_HAVE_XXX and let the team here know about this.

We currently use full_mips-eng internally and also test our sources on MIPS emulator. MIPS support for the emulator is now uploaded. I still need to upload the build scripts.
 

>This will ensure that platform modifications that break the >MIPS-dependent build could be caught reasonably soon.

Sure. This is something MIPS team is looking forward to.

Bionic libc comments:

I will try to workaround some of the ifdefs __mips__ in C library definitions and submit them as smaller patches encompassing a file or couple. Some of them which are due to kernel differences between MIPS and other architectures might have to stay.

Thanks,
Raghu

Patch-set: 1
Reviewer: Raghu Gandham <1003296@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0