Update patch set 2
Patch Set 2:

Well, no. The existing code does handle the problem. It handles it by looking to see if there is a default route for the protocol(s) that are being tried.

The existing code does have issue #1, but I don't really see that as an issue, since applications must be prepared for the fact that new DNS records may show up and be returned to them at any time.

I believe that applications on secondary connections are already taken care of because they use their own per-app DNS servers (there's a check in the code for that). So you don't need to worry about #2.

What if you just made _have_ipv4() return false if the IPv4 default route was via a transition method?

Patch-set: 2