Update patch set 2
Patch Set 2:

Another thing we could do is use another property such as "net.ipv4.compat". That property would be set if the IPv4 default route were over a transition mechanism. It could be set on IPv6-only wifi if we support that in the future, and so on.

Patch-set: 2