Update patch set 1
Patch Set 1:

No, that is not from the upstream. We found that bug and fixed it.
Per your comment, we searched the latest FreeBSD libm implementation, and found a similar fix there: 
http://svnweb.freebsd.org/base?view=revision&revision=176101

Do you want to pick up this patch from freebsd?
If yes, will you do by yourself or I re-submit a patch?
If no, Or you will upgrade the android libm to the last upstream version?

Patch-set: 1