Update patch set 1
Patch Set 1:

+1 to what Bruce said.

i've received the mail with the tests. i really need to find the time to start using external/gtest/ to write some bionic/ tests. we should pull back and translate the old tests from system/ or wherever they are at the moment. in the meantime, if you just upload a patch to gerrit with the tests, at least they'll be visible. (don't worry about whether they actually build and run if you don't have time either. since the state of bionic testing is currently so abysmal it would be unreasonable to _insist_ on unit tests. but that's definitely where i'd like us to end up in the long run, especially for pure functions.)

Patch-set: 1