Update patch set 2
Patch Set 2:

Google people, any thoughts about this change?

I'm in the process of verifying that this fixes the boot deadlocks that we've had for the last 2-3 weeks, and if that indeed appears to fix things I'd like to proceed to submit this change (or revert the original change, which doesn't please me that much).

Patch-set: 2
Label: Code-Review=0