Update patch set 2
Patch Set 2:

Refine the patch comment to make it easier to understand.

This patch is just for RFC, please don't merge it so far. We guess we found one bug in bionic regarding to the __cxa_guard_acquire/__cxa_guard_release. Please read the patch for more details.

Patch-set: 2
Label: Code-Review=0