Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

i think the original change was wrong, but this change is wrong too. we should choose between kill(getpid(), sig) and pthread_kill(pthread_self(), sig), like POSIX says. the current code is incorrect because it's using a tid as a pid. the new code would be incorrect because it doesn't include a tgid.

can we get away with always calling pthread_kill(pthread_self(), sig), i wonder? or are we used in contexts where that won't work?

Patch-set: 1