Update patch set 1
Patch Set 1:

Using tkill() for raise() is not that bad.

From tkill man page:

"tkill() is an obsolete predecessor to tgkill(). It only allows the target thread ID to be specified, which may result in the wrong thread being signaled if a thread terminates and its thread ID is recycled. Avoid using this system call."

The danger in using tkill() is when the thread ID is recycled, but this will not happen in raise(), when the calling thread will be alive when it sends a signal synchronously to self. 

Using pthread_kill(pthread_self(), signo) is not much better, which uses tkill(), too.
Since bionic libc doesn't have a tgkill() system call wrapper, using tkill() in raise() is a simple and working solution.

Patch-set: 1
Reviewer: Cervantes Yu <1011386@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0