Update patch set 1
Patch Set 1:

Elliott,
We asked the question on FreeBSD mail list. You could check it here: 
http://lists.freebsd.org/pipermail/freebsd-threads/2012-April/005273.html

I suppose it could be an issue on the platform based on ARM (Could not be an issue on x86 because x86 has more strict memory model).

https://android-review.googlesource.com/#/c/35380/ could be dropped because the failure patch is not possible to happen.

https://android-review.googlesource.com/#/c/35371/ should be valid change. The Android __sinit has different code with the one in FreeBSD. But the SMP safe issue is there for ARM.

Let us know if you have further comment.

Patch-set: 1