Update patch set 1
Patch Set 1:

Hi Elliott,
https://android-review.googlesource.com/#/c/35370/ was merged.
I suppose this one should be merged as well.

The bionic __sinit implementation is not exactly same as FreeBSD implementation. In bionic, __sinit is typical singleton pattern. And need memory barrier to make sure SMP safe.

And again, the issue may not happen on x86. But it could happen on ARM/MIPS in theory.

Patch-set: 1