Update patch set 1
Patch Set 1:

Hell, I'm very late to the party :-)

I think this problem never occurs on x86 because on this platform, all loads have acquire semantics, and all store release semantics. However, this is a very nice find for a real problem on ARM and other architecture.

However, I would prefer a slightly different solution, where the actual release operation is in findfp.c itself (e.g. in a new function like __sfp_free), which would perform the store-release operation. I just don't like the idea of random memory barrier operations scattered around separate source files (the patch is completely correct however).

Another point of discussion, that is not strictly related to this patch is the fact that we should really avoid using full memory barriers on x86 when they are not needed (ANDROID_MEMBAR_FULL is costly and unnecessary here on x86)

This means it'd be a good idea to implement __bionic_atomic_load_acquire() + __bionic_atomic_store_release() in libc/private/bionic_atomic_inline.h, then replace our use of ANDROID_MEMBAR_FULL in the C library with these.

Unfortunately, the version in the internal Android tree of this header is now very different from the one in AOSP. I think it'd be a good idea to back-port them first to AOSP, then implement this improvement on top of that.

Elliot, what do you think about this?

Patch-set: 1
Label: Code-Review=0