Update patch set 1
Patch Set 1:

Hi Turner,
Thanks a lot for you feedback.

For the issue on x86. Yes. It will not be an issue on Intel chip. But none-intel x86 could have out of order STORE operation, it could be possible that the issue happens. So we'd better to add the memory barrier operation here and put different implementation to arch specific related code. Let's say, we could define the memory barrier operation as compiler barrier for Intel chip (and a real memory barrier operation for ARM and mips).

For atomic APIs, I suppose they will be implemented using prefix "lock". Unfortunately, prefix "lock" contain a full memory barrier on x86. So it could add the additional overhead which is not necessary.

Patch-set: 1