Update patch set 2
Patch Set 2:

Yes, proper overflow handling would be nice. Since we use strtoll for parsing that, we can't handle all the cases a "real" strtod would handle though, so we'd overflow in more cases than a fullblown implementation.

I guess that can be added on top as a separate change though - this one at least handles quite a bit of the valid cases.

Who can approve it, Elliott?

Patch-set: 2