Update patch set 1
Patch Set 1:

i can read the comments :-) my question was really why you chose openbsd for anything, and why you chose netbsd over freebsd. going forward i'd like to get back into a position where we track upstream, and as part of that i'd like to reduce the number of upstreams we have. ditching openbsd seems like a no-brainer, but the choice between netbsd and freebsd is much less obvious to me.

what i'd really like is an upstream that might be interested in taking our changes. we have open bugs in stuff like getopt(3), for example, that are still broken in all the upstreams, and i'd like to get to a position where we can fix these bugs in the right place.

Patch-set: 1