Update patch set 5
Patch Set 5: I would prefer that you didn't submit this

"Currently fork() is really unusable in multithreaded programs."

yes, but this has always been true. "don't do that." or, more specifically, "don't call into dlmalloc in the child."

i don't want us to be the first to take an experimental (and to my mind misguided) change to dlmalloc.

Patch-set: 5