Update patch set 1
Patch Set 1:

MIPS changes were picked up mostly from OpenBSD based on the comments in the corresponding ARM files. However, if we need to pick the files again, NetBSD is the preferred one as compared to freeBSD. I will resubmit the patch after rebasing it to the current master.

Patch-set: 1