Update patch set 2
Patch Set 2:

can you put a canonical copy somewhere on the AOSP website? if that's supposed to be the preferred form, it's actually one of the least common ones, and before i start telling people they're doing it wrong, i'd like to have somewhere authoritative to point them to!

(i also don't understand when we choose to use the BSD blurb versus the APL blurb.)

Patch-set: 2