Update patch set 2
Patch Set 2: (2 inline comments)

I understand that this is a controversial change, but the main concern is that code that works correctly on x86 or mips can cause crashes on the ARM in a way that is hard to spot and/or debug (at program exit time).

I can rework the patch to not change the behavior but issue a warning, if you prefer.

Patch-set: 2
Label: Code-Review=0