Update patch set 1
Patch Set 1:

but ffs is a function, and callers should be able to #undef ffs and still call ffs.

maybe have both use an inline ffs that calls __builtin_ffs?

but there's a *shitload* of these functions in gcc, and i can't say i'm particularly excited about going through this for every single one of them. (and, on the other hand, if ffs is the only one where clang insists on the __builtin_ prefix, that again makes me think that's a clang bug.)

Patch-set: 1