Update patch set 2
Patch Set 2: No score

(13 inline comments)

> Is the code size of "lea constant(reg), reg" less of an issue than the scheduling improvement over "add constant, reg" where this pattern is used?

I use lea instead of add according to Software Optimization Guide (atom section).(LEA should not be used in the technique of adding two values and produce the result in a third register. LEA should be used for address computation.)
My case is address computation.


You are right, there is a bug in wmemcmp which can be reproduced just for N=1,2,3,4 and equal strings.
It's simple to fix it if load to other register instead of eax, ecx will be fine. Thanks.

Patch-set: 2
Label: Code-Review=0