Update patch set 2
Patch Set 2:

Thanks Dmitri, when it comes to the hand scheduling and branch fusion as this patch is coming from Intel we'll trust you to know what works best on Atom. I was using my rusty memories of the optimization manual. Clearly it makes sense to merge the patch with wmemcmp fixed. Thanks again.

Patch-set: 2