Update patch set 1
Patch Set 1:

Reading through this again, isn't the correct solution to specify -std=c89 in the CFLAGS? We don't want C99 behavior if the code was intended for C89. That likely makes the warnings go away and won't impact the current gcc compiler either (it's a nop since C89 is the default for gcc).

Patch-set: 1
Label: Code-Review=0