Update patch set 5
Patch Set 5:

Just adding my 2c here:

According to the quoted spec, two things are still missing from the way R_ARM_COPY_RELOC entries are handled:
- the linker should check whether the size of the target area is sufficient to contain the copy (this can be addressed trivially by taking the size of the local reference and comparing it to the non-local one);
- ALL OTHER references should be pre-empted and point to the copy after relocation.

In terms of the example, it means the expected output should be something like

global = 0x42 (0x11000)
global = 0x42 (0x11000)

which indicates that the reference inside the .so now also (correctly) points to the copy.

The latter is particularly hairy, as it requires another traversal of all needed libraries to fixup references that have already been relocated. It also means that copy relocations are likely incompatible with shared libraries built with -Bsymbolic (which is the default on Android) as it forces global references to be bound to definitions inside the library.

Patch-set: 5