Update patch set 5
Patch Set 5:

Using this change is like putting a napkin over an open can of worms. You should not depend on being able to do much other than exec* after fork from a multithreaded process. It may work now, but that guarantee is hard to keep when any new synchronization introduced in the future may deadlock. The correct approach is to change the app to not use fork.

Can you elaborate on the particular requirement for WebGL?

Patch-set: 5