Update patch set 1
Patch Set 1:

At least on my device, when I revert the libcore change and add a log line, I see that the mask is 031

My code:

static jint Posix_umask(JNIEnv*, jint mask) {
    ALOGE("mask=%x, called_mask=%x\n", mask, (mode_t) mask);
    return umask(mask);
}

And the logging output:

E/Posix   (  400): mask=93e00019, called_mask=19

The low order bits always seem to be 0x19 (031), which is always in the range 0 to 0777.

So this change wouldn't have detected this bug.

Patch-set: 1