Update patch set 1
Patch Set 1:

the various BSDs have __unused in the same header, and it's been in the NDK since the beginning. there's a ton of cruft in <sys/cdefs.h> that i wish wasn't public, but it seems odd to clean up __unused without looking at the other cruft, and that seems like a big job to take on right now.

the whole purpose of this script is rewriting these files anyway :-)

Patch-set: 1