Update patch set 2
Patch Set 2:

In general, no, the use of aapcs-vfp here is no different from building hard-float calling convention in with function resolution rather than compilation unit resolution (in this case of course only non-API functions).  So to answer your question (if I understand it correctly) it's not target-specific, but we use target-specific #define's since the aapcs-vfp changes are sort of baked in with some fairly target-specific optimizations (the pow() changes will likely only help fairly superscalar out-of-order processors like Krait/A15 but not so much for Scorpion/A9).  It would also be possible to separate out the aapcs-vfp changes and conditionally include them for anything that supports hard-float but the performance improvement of that alone without any of the algorithmic changes might be a hard sell.

Patch-set: 2
Label: Code-Review=0