Update patch set 1
Patch Set 1:

bdc, kroot, nnk: note that glibc doesn't include these. glibc has a linux/keyctl.h containing constants, but doesn't declare the functions. the implementations are in a separate libkeyutils and there's a keyutils.h that comes with that library. do we plan on using the kernel's key management stuff in Android? if so, it seems reasonable to add this to bionic. if not, we should probably just let users use syscall(3) themselves instead.

Patch-set: 1