Update patch set 4
Patch Set 4:

Thanks Iliyan, in this change the regions are labelled "libc malloc". What you see in /proc/pid/maps is "/dev/ashmem/libc malloc" so this shouldn't conflict with tools trying to find "[heap]". We could name these allocations "[heap]" but this would conflict with the brk region at the end of data. Enh's proposal was to name them "[large heap]" but that's not quite true either as mmap is used as a fallback by malloc when brk fails (so you end up with small page sized regions labelled "[large heap]"). Given this, and that its an easy change, "libc malloc" strikes me like the least worst solution for the name - and without the square braces as we're not part of the kernel.

Patch-set: 4