Update patch set 1
Patch Set 1:

Henrik, I don't quite understand how you are measuring different values after my change. The code I am removing is not affected by either of the defines you mentioned. Specifically, the defines you mention are only used if NEON_UNALIGNED_ACCESS is defined, and my change is in the code if this is not defined. I am not removing the arm loop from when NEON_UNLAIGNED_ACCESS is defined, so it should not affect the performance on a galaxy nexus (and I measured that it does not).

Also, can you tell me what tool you are using to measure the changes? I tested this change on three different platforms (galaxy nexus, nexus 7, and nexus 10). But I didn't go down to 1K size, I tested on 8KB, 16KB, 32KB, 64KB, 128KB, 256KB, 512KB, 1MB, 2MB sizes. I was using the micro_bench tool with some slight modifications (priority of process elevated, locked to a single core, and I added a computation of standard deviations). For the nexus 10, the performance after my change is about 10000 MB/s, with the code as submitted it's about 5000MB/s.

One other quick thing, you mention that NEON_MEMCPY_ALIGNMENT_DIVIDER should be set to 224 and that  NEON_MEMSET_DIVIDER should be set to 132. Where is that supposed to happen, in the device Makefile file or somewhere else? I searched the tree a few times and those values are never set so maybe something got merged incorrectly.

Patch-set: 1