Update patch set 1
Patch Set 1:

I just realized that I didn't look at the galaxy nexus results properly. The current code has a speed of about 4000MB/s, while the fix and the original code has a speed of about 2350MB/s. I did a quick mod of the arm-only code for a 64 byte cache and I did see a speed improvement on nexus 10, but only about 500MB/s (goes from 5100MB/s to 5650MB/s).

However, doing a memcpy where src and dst are not aligned on a galaxy nexus yields the original slow speed which implies that it would be better to have a single routine that aligned src and dst since the mainloop speed up is so much faster with the arm-only code.

I'm going to change this code just a bit so that we still get the benefit of the galaxy nexus change and fix the regression on the nexus 10.

Patch-set: 1