Update patch set 1
Patch Set 1:

In answer to Christopher Ferris:

I didn't measure this exact patch. I took the current AOSP version and added the 'Neon unaligned' flags and 'memcpy_alignment' value, since they should have been set (in the makefile). I then measured with and without the branch to the ARM-only code.
Enabling the neon unaligned flag should give a boost to smaller unaligned memcpy operations which are quite common.

I measured with a small tool that I wrote myself. I picked the results from 1K size simply because it's a common size used in memcpy calls. The most common is 1 byte, but that generates too much overhead to measure consistently. Sizes above 1024K is very uncommon in normal apps, but sadly common in benchmark apps.
8KB is good for measuring cached performance, but unlike benchmark tests, normal apps don't copy the same data to the same destination more than once. That's why I also measure 'random' memcpy operations. To get the 'real' performance values.
Good to hear that Neon is always faster than the ARM-core when reading and writing lots of memory on A15, even for aligned operations.
Sadly, the standard A9 is not. That's why I added an ARM-only version for when the source and destination are aligned.

Yes, NEON_MEMCPY_ALIGNMENT_DIVIDER and NEON_MEMSET_DIVIDER should be set in the makefile. I set them based on similarly named build flags in the BoardConfig.mk, so they can be configured on a product basis. I don't know why this wasn't merged by us (STE). I had a discussion about it with Elliott a while back, and about the NEON_UNALIGNED-flag. 
But this may be solved in a different way if, as Ben says, you are going to have multiple versions of memcpy, for different processor generations, soon.

Patch-set: 1