Update patch set 1
Patch Set 1:

I thought I was going to find some way to put the code in a define that was only available on nexus10, but no such luck. I'm going to drop this change in favor of the new structure to support different memcpy implementations.

I'll leave this up for a bit before removing it.

Patch-set: 1
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0