Update patch set 3
Patch Set 3:

How about we just make raise() a bit smarter?
I just tried this and it fixes all of the sigaction()
LTP test and it makes sense to me that we should call
kill() if the program is single threaded and comparing
the tid to the pid is a valid way to do that; thought 
perhaps not the fastest.

Here's the version I just tried. Hope it looks on when you see it:
-----------------------------------------------------------
 #include <pthread.h>
+#include <sys/types.h>
+#include <unistd.h>
 
 int raise(int sig) {
-  int rc = pthread_kill(pthread_self(), sig);
+  int rc;
+  int tid = gettid();
+  int pid = getpid();
+
+  if (tid == pid) {
+    rc = kill(pid, sig);
+    return rc;
+  }
+
+  rc = pthread_kill(pthread_self(), sig);
   if (rc != 0) {
     errno = rc;
     return -1;
   }
   return 0;
------------------------------------------------------------

-piet

Patch-set: 3