Update patch set 1
Patch Set 1:

Ok, I just gave a try.
It fails because lots of variables in $(BUILD_CTS_EXECUTABLE) are not defined yet at this point. However BUILD_CTS_EXECUTABLE itself is indeed available at this point. That seems a bug to me.
So the question for the CTS guys is: should we allow to define cts tests in directories outside the cts git project? 
If so, we should fix the bug and make sure "include $(BUILD_CTS_EXECUTABLE)" works everywhere; If not, we should make BUILD_CTS_EXECUTABLE not available at all outside the cts project.

Patch-set: 1